Add Terrain Painting based on Biome

+++ Assets/scripts/TexturePainting/BaseTexturePainter.cs
@@ -37,7 +37,7 @@ public class BaseTexturePainter : MonoBehaviour
     [SerializeField] [Range(0f, 1f)] protected float Strength = 1f;
     
     public virtual void Execute(TexturePainting painter, int mapResolution, float[,] heightMap, Vector3 heightmapScale, float[,] slopeMap, 
-                                float[,,] alphaMaps, int alphaMapResolution, byte[,] biomeMap = null, int biomeIndex = -1, BiomesConfig biome = null)
+                                float[,,] alphaMaps, int alphaMapResolution, int[,] biomeMap = null, int biomeIndex = -1, BiomesConfig biome = null)
     {
         Debug.LogError("No implementation of Execute function for " + gameObject.name);
     }
