+++ Assets/scripts/TexturePainting/TexturePainter_Random.cs
@@ -18,7 +18,7 @@ public class TexturePainter_Random : BaseTexturePainter
     [SerializeField] List<RandomPainterConfig> PaintingConfigs;
 
     public override void Execute(TexturePainting painter, int mapResolution, float[,] heightMap, Vector3 heightmapScale, float[,] slopeMap, 
-                            float[,,] alphaMaps, int alphaMapResolution, byte[,] biomeMap = null, int biomeIndex = -1, BiomesConfig biome = null)
+                            float[,,] alphaMaps, int alphaMapResolution, int[,] biomeMap = null, int biomeIndex = -1, BiomesConfig biome = null)
     {
         int baseTextureLayer = painter.GetLayerForTexture(BaseTexture);
